pip_requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-appconf==1.1.0
django-compressor==4.5.1
django-htmlmin==0.11.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
html5lib==1.1
idna==3.10
pillow==11.3.0
rcssmin==1.1.2
requests==2.32.5
rjsmin==1.2.2
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
webencodings==0.5.1
